removes redundancy in tests, deletes DataProviderIbanTest.java
