+++ iban-validator/src/test/java/com/intive/iban/validator/IbanValidatorTest.java
@@ -1,5 +1,6 @@
 package com.intive.iban.validator;
 
+import com.tngtech.java.junit.dataprovider.DataProvider;
 import com.tngtech.java.junit.dataprovider.DataProviderRunner;
 import com.tngtech.java.junit.dataprovider.UseDataProvider;
 import org.junit.Assert;
@@ -25,30 +26,36 @@ public class IbanValidatorTest {
         this.validator = factory.getValidator();
     }
 
+    @DataProvider
+    public static String[] dataProviderValidIban() {
+        return new String[]{
+                "60102010260000042270201111",
+                " 61109010140000071219812874",
+                "27114020040000300201355387 "};
+    }
+
+    @DataProvider
+    public static String[] dataProviderInvalidIban() {
+        return new String[]{
+                "601020102600000422=70201111",
+                "2 7114020040000300201355387",
+                "\"61109010140000071219811\\\\n\"",
+                "60-1020-1026-0000-0422-7020-1111",
+                "",
+                " ",
+                null};
+    }
+
     @Test
-    @UseDataProvider(value = "dataProviderValidIban", location = DataProviderIbanTest.class)
+    @UseDataProvider("dataProviderValidIban")
     public void testValidIban(String ibanData) {
         SampleIban iban = new SampleIban(ibanData);
         Assert.assertTrue(iban.isIbanValid());
     }
 
     @Test
-    @UseDataProvider(value = "dataProviderInvalidIbanBadCountryCodeOrLength", location = DataProviderIbanTest.class)
+    @UseDataProvider("dataProviderInvalidIban")
     public void testInvalidIbanBadCountryCodeOrLength(String ibanData) {
-        SampleIban iban = new SampleIban(ibanData);
-        Assert.assertFalse(iban.isIbanValid());
-    }
-
-    @Test
-    @UseDataProvider(value = "dataProviderInvalidIbanBadFormat", location = DataProviderIbanTest.class)
-    public void testInvalidIbanBadFormat(String ibanData) {
-        SampleIban iban = new SampleIban(ibanData);
-        Assert.assertFalse(iban.isIbanValid());
-    }
-
-    @Test
-    @UseDataProvider(value = "dataProviderInvalidIbanUnallowedCharacters", location = DataProviderIbanTest.class)
-    public void testInvalidIbanUnallowedCharacters(String ibanData) {
         SampleIban iban = new SampleIban(ibanData);
         Assert.assertFalse(iban.isIbanValid());
     }
